+++ test/Makefile.am
@@ -146,6 +146,7 @@ common_sources = \
 	../common/Unit.cpp \
 	../common/FileUtil.cpp \
 	../common/Util.cpp \
+	../common/Util-desktop.cpp \
 	../common/StringVector.cpp \
 	../common/TraceEvent.cpp \
 	../common/Simd.cpp \
@@ -178,7 +179,7 @@ unit_base_la_LIBADD += -lssl -lcrypto
 endif
 
 fakesockettest_CPPFLAGS = -g
-fakesockettest_SOURCES = fakesockettest.cpp  ../net/FakeSocket.cpp ../common/DummyTraceEventEmitter.cpp ../common/Log.cpp ../common/Util.cpp
+fakesockettest_SOURCES = fakesockettest.cpp  ../net/FakeSocket.cpp ../common/DummyTraceEventEmitter.cpp ../common/Log.cpp ../common/Util.cpp ../common/Util-desktop.cpp
 fakesockettest_LDADD = $(CPPUNIT_LIBS)
 
 # old-style unit tests - bootstrapped via UnitClient
